minor update_checkstyle clean
