+++ core/src/main/java/greencity/controller/OrderController.java
@@ -25,6 +25,7 @@ import greencity.dto.user.UserVO;
 import greencity.entity.user.User;
 import greencity.enums.OrderStatus;
 import greencity.enums.PaymentStatus;
+import greencity.exceptions.NotFoundException;
 import greencity.service.ubs.NotificationService;
 import greencity.service.ubs.UBSClientService;
 import greencity.service.ubs.UBSManagementService;
@@ -33,6 +34,7 @@ import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.validation.annotation.Validated;
@@ -433,5 +435,27 @@ public class OrderController {
     @GetMapping("/orders/{id}/tariff")
     public ResponseEntity<TariffsForLocationDto> getTariffForOrder(@PathVariable Long id) {
         return ResponseEntity.status(HttpStatus.OK).body(ubsClientService.getTariffForOrder(id));
+    }
+
+    /**
+     * Check if a tariff exists by its ID.
+     *
+     * @param id The ID of the tariff to check.
+     * @return ResponseEntity with a boolean indicating whether the tariff exists.
+     * @throws NotFoundException if the tariff with the specified ID is not found.
+     * @author Yurii Ososvskyi
+     */
+    @ApiOperation(value = "Check if tariff exists by Id")
+    @ApiResponses(value = {
+        @ApiResponse(code = 200, message = HttpStatuses.OK),
+        @ApiResponse(code = 400, message = HttpStatuses.BAD_REQUEST),
+        @ApiResponse(code = 401, message = HttpStatuses.UNAUTHORIZED),
+        @ApiResponse(code = 403, message = HttpStatuses.FORBIDDEN),
+        @ApiResponse(code = 404, message = HttpStatuses.NOT_FOUND)
+    })
+    @GetMapping(value = "/check-if-tariff-exists/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<Boolean> checkIfTariffExistsById(@PathVariable Long id) {
+        Boolean exists = ubsClientService.checkIfTariffExistsById(id);
+        return ResponseEntity.status(HttpStatus.OK).body(exists);
     }
 }
